@@ -51,8 +51,8 @@
     "optional-require": "^1.0.2",
     "require-at": "^1.0.6",
     "url-loader": "^4.1.0",
-    "webpack": "^5.24.3",
-    "webpack-cli": "^4.5.0",
+    "webpack": "^5.33.2",
+    "webpack-cli": "^4.6.0",
     "webpack-config-composer": "^1.1.3",
     "webpack-stats-plugin": "^1.0.3",
     "xsh": "^0.4.5"
