@@ -2,26 +2,27 @@
 	MIT License http://www.opensource.org/licenses/mit-license.php
 	Author Tobias Koppers @sokra and Zackary Jackson @ScriptedAlchemy
 */
+/* eslint-disable */
 
-"use strict";
+// const { validate } = require("schema-utils");
+const schema = require("webpack/schemas/plugins/container/ModuleFederationPlugin.json");
+const SharePlugin = require("webpack/lib/sharing/SharePlugin");
+const ContainerReferencePlugin = require("webpack/lib/container/ContainerReferencePlugin");
 
-const { validate } = require("schema-utils");
-const schema = require("../../schemas/plugins/container/ModuleFederationPlugin.json");
-const SharePlugin = require("../sharing/SharePlugin");
-const ContainerPlugin = require("./ContainerPlugin");
-const ContainerReferencePlugin = require("./ContainerReferencePlugin");
+import { ContainerPlugin } from "./ContainerPlugin";
 
 /** @typedef {import("../../declarations/plugins/container/ModuleFederationPlugin").ExternalsType} ExternalsType */
 /** @typedef {import("../../declarations/plugins/container/ModuleFederationPlugin").ModuleFederationPluginOptions} ModuleFederationPluginOptions */
 /** @typedef {import("../../declarations/plugins/container/ModuleFederationPlugin").Shared} Shared */
 /** @typedef {import("../Compiler")} Compiler */
 
-class ModuleFederationPlugin {
+export class ModuleFederationPlugin {
+  _options: any;
   /**
    * @param {ModuleFederationPluginOptions} options options
    */
   constructor(options) {
-    validate(schema, options, { name: "Module Federation Plugin" });
+    // validate(schema, options, { name: "Module Federation Plugin" });
 
     this._options = options;
   }
@@ -33,15 +34,19 @@ class ModuleFederationPlugin {
    */
   apply(compiler) {
     const { _options: options } = this;
+
     const library = options.library || { type: "var", name: options.name };
+
     const remoteType =
       options.remoteType ||
       (options.library && schema.definitions.ExternalsType.enum.includes(options.library.type)
         ? /** @type {ExternalsType} */ options.library.type
         : "script");
+
     if (library && !compiler.options.output.enabledLibraryTypes.includes(library.type)) {
       compiler.options.output.enabledLibraryTypes.push(library.type);
     }
+
     compiler.hooks.afterPlugins.tap("ModuleFederationPlugin", () => {
       if (
         options.exposes &&
@@ -51,6 +56,7 @@ class ModuleFederationPlugin {
       ) {
         new ContainerPlugin({
           name: options.name,
+          entry: options.entry,
           library,
           filename: options.filename,
           exposes: options.exposes
@@ -76,5 +82,3 @@ class ModuleFederationPlugin {
     });
   }
 }
-
-module.exports = ModuleFederationPlugin;
