@@ -10,7 +10,7 @@ async function loadRemoteModule(scope, module) {
   const container = window[scope]; // or get the container somewhere else
   // Initialize the container, it may provide shared modules
   await container.init(__webpack_share_scopes__.default);
-  const factory = await window[scope].get(module);
+  const factory = await window[scope].get(module, scope);
   const Module = factory();
   return Module;
 }
